Moving spark reproduction from ingest to reproject package
Mirrors the structure in raster project
